[flue] Extract style-guide fan-out mechanics into lib/style-guide-fanout.ts
